refactor: 리덕스 툴킷 스토어 부분 폴더추가,정리
