+++ src/store/user/userApi.js
@@ -1,0 +1,3 @@
+export function fetchUser() {
+  return new Promise(resolve => setTimeout(() => resolve(), 500));
+}
